setup : adjusted tests naming & structure - added --exclude-archs to test script for faster build times - main.js consolidates tests - moved tests from main.js to app.js - renamed tests/array.test.js to tests/array.js
