+++ package.json
@@ -3,8 +3,8 @@
   "private": true,
   "scripts": {
     "start": "meteor run --exclude-archs web.browser.legacy,web.cordova",
-    "test": "meteor test --once --driver-package meteortesting:mocha",
-    "test-app": "TEST_WATCH=1 meteor test --full-app --driver-package meteortesting:mocha",
+    "test": "meteor test --exclude-archs web.browser.legacy,web.cordova --once --driver-package meteortesting:mocha",
+    "test-app": "TEST_WATCH=1 meteor test --exclude-archs web.browser.legacy,web.cordova --full-app --driver-package meteortesting:mocha",
     "visualize": "meteor --production --extra-packages bundle-visualizer",
     "lint": "eslint --no-warn-ignored client/ imports/ server/",
     "lint:fix": "eslint --fix --no-warn-ignored client/ imports/ server/",
