+++ tests/app.js
@@ -1,0 +1,20 @@
+import assert from 'assert';
+
+describe('meteor-app', function () {
+  it('package.json has correct name', async function () {
+    const { name } = await import('../package.json');
+    assert.strictEqual(name, 'SkillTree');
+  });
+
+  if (Meteor.isClient) {
+    it('client is not server', function () {
+      assert.strictEqual(Meteor.isServer, false);
+    });
+  }
+
+  if (Meteor.isServer) {
+    it('server is not client', function () {
+      assert.strictEqual(Meteor.isClient, false);
+    });
+  }
+});
