+++ tests/main.js
@@ -1,21 +1,3 @@
-import assert from "assert";
-import './array.test.js';
-
-describe("meteor-app", function () {
-  it("package.json has correct name", async function () {
-    const { name } = await import("../package.json");
-    assert.strictEqual(name, "SkillTree");
-  });
-
-  if (Meteor.isClient) {
-    it("client is not server", function () {
-      assert.strictEqual(Meteor.isServer, false);
-    });
-  }
-
-  if (Meteor.isServer) {
-    it("server is not client", function () {
-      assert.strictEqual(Meteor.isClient, false);
-    });
-  }
-});
+// Tests
+import '/tests/app';
+import '/tests/array';
